### Brief description of the changes:
2 more steps of optimizing `biter_raffle`:
1) Switch `raffle` table into array - 104ms => 84ms - ~20% improvement
2) Reuse one "global" `raffle` table instead recreating it on each call - 84ms => 64ms - another ~20% improvement

Unfortunately, switching to arrays broke the tests, so they had to be updated as well.
It probably can be optimized a bit more, by skipping unnecessary `return` statements, but for sake of comparability (and in order not to break any more tests), I've left it as is for now

Based on ideas form @vadcx  in #477 

### Tested Changes:
- [x] I've tested the changes locally.
- [ ] I've not tested the changes.